Add some missing dependencies
